refactor: simplify FilteredRepository abstraction
