+++ packages/filtered_repository/lib/src/filtered_repository.dart
@@ -1,14 +1,16 @@
 import 'package:filtered_repository/src/models/models.dart';
 
 /// An interface for a SpaceX repository using filters.
-abstract class FilteredRepository<T, S extends Feature> {
-  /// An abstract constructor of [FilteredRepository].
+abstract class FilteredRepository<T, S extends SortingParameter,
+    R extends Filtering> {
+  /// An abstract constructor of a [FilteredRepository].
   const FilteredRepository();
 
-  /// Fetches specific SpaceX objects that meet the specified [parameters].
+  /// Fetches specific SpaceX objects that meet the specified parameters.
   Future<List<T>> fetchFiltered({
     required int amount,
     required int pageNumber,
-    FilterParameters<S> parameters,
+    Sorting<S>? sorting,
+    R? filtering,
   });
 }
