--- packages/filtered_repository/lib/src/models/feature.dart
@@ -1,7 +1,0 @@
-// ignore_for_file: one_member_abstracts
-
-/// A feature of an object.
-abstract class Feature {
-  /// Converts this feature to object's field name in `snake_case`.
-  String toFieldName();
-}
